requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
celery==4.4.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
environ==1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==4.6.11
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==20.9
prompt-toolkit==3.0.19
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.17.1
python-http-client==3.3.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
uritemplate==3.0.1
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
